Backport no argument reordering as an option
In the unreleased version 2, the argument reordering has been removed
(in f585ec7cb8b0d9189ac9016fe83f911f3a360f2e) since it only worked if
all of the arguments appeared before all of the flags, but not if they
were intermixed which was of limited utility and caused some confusion.
This commit allows enabling of this future behavior via SkipArgReorder.

Ideally we'd support complete intermingling of flags and arguments, but
this is unlikely to happen until we switch flag parsers.

Fixes #515