Allow Ignoring Usage Errors
In some rare cases, it's handy to allow the application to handle
use-case errors in other ways, without requiring an error to be
returned. This allows the application written to handle the usage
error in other ways. Or, allow the application to pass the flags
on to subsequent child processes regardless of the host application's
flag parsing.

```
$ go test -run="TestApp_IgnoreOnUsageError"
PASS
ok      github.com/grubernaut/cli       0.001s
```