Allow zsh completion to be distributed like bash
The readme describes a distribution pattern for bash_autocomplete. In
this pattern, the file is renamed to match the name of the program for
which the  autocompletion support is being added. When following this
pattern, it is not necessary to set the PROG environment variable.

Update zsh_autocomplete so that it can be distributed the same way.