ci: add go 1.17.x and 1.18.x, fix failing test
One test is failing with Go 1.17+ since it now panics instead of
returning an error. Catch the panic and consider the test passed.

With that out of the door, add Go 1.17.x and 1.18.x to the testing
matrix.

While at it, do not use && in run since it is executed by a shell with
set -e, thus making && redundant.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>