use correct context in After function with subcommand
When adding values to a context in a Before function that value is not
present in the After function when a subcommand was run. Without the
subcommand the value is present.

This fixes that be returning the new context from the subcommand and
changing the context in the calling function. Defer'd function now pick
up the new context.

Fixes #2098