feat: allow the user to stop processing flags after seeing N args
We have switched from v2 to v3 mainly because we want to support parsing flags
after the arguments. There are many other benefits, but this one drove the
work.

We do have some commands where the old v2 behaviour is desirable, specifically,
when we pass arguments through. For example our ssh command:

  tool ssh MACHINE-NAME ls -la

To support this on a per-command basis, added an option to the Command struct
called StopOnNthArg. It terminates flag parsing like '--'.