fix completion of "--"
Use https://cli.urfave.org/v3/examples/completions/shell-completions as a
reference for completions.

The previous version completed "-" -> "--" -> "--help,", but the added comma was
incorrect, as no such flag exists and the tool does not accept it. A similar
issue occurred with "--version", which was shown as "--version,".

The root cause was that completion was explicitly disabled for this case:

$ greet -- --generate-shell-completion

As a result, it printed the default help message:

> NAME:
>    greet - A new cli application
>
> USAGE:
>    greet [global options] [command [command options]]
>
> COMMANDS:
>    add, a       add a task to the list
>    complete, c  complete a task on the list
>    template, t  options for task templates
>    help, h      Shows a list of commands or help for one command
>
> GLOBAL OPTIONS:
>    --help, -h  show help

Bash used "--help," as a flag suggestion, which was incorrect.

In this commit, completion is enabled for the "-- --generate-shell-completion"
case, and a test is added to verify correct behavior.

With this change, both "-" and "--" now complete to "--help", which is correct.